# Optimize Burundi observed holidays