Performance tests:
```
def calc_hol1():
    for year in range(1962, 2100):
        hols = Burundi(years=year)

def calc_hol2():
    hols = Burundi(years=range(1962, 2100))

t1 = timeit.timeit(calc_hol1, number=100)
t2 = timeit.timeit(calc_hol2, number=100)
```
| Test | Current beta |   PR  |
|:----:|:------------:|:-----:|
|  t1  |        3.114 | 3.117 |
|  t2  |        6.797 | 2.920 |